Update RCDirectionPanel.py
This command works with the new RCClass added. However because the panels are in their own classes the motor speed variables do no communicate with one-another and so when the speed is set in the speed panel it is not picked up by the directional panel. Whether this can be solved by changing var's too global or whether I need to rejig some methods, I'm not sure.