RCPassword
Quick little security password for my RC, thought I might as well use the touch sensors on the explorerhat as they make a great PIN entry system. Found a majority of this programming online but added the ability to change the password, next step will be saving and loading a password in. Might add it into my settings folder but I'm not sure how without deleting other settings.